fix: get websocket options working
